5.1.14 Разделяет utils на common и render, переподключает render.js
